+++ NewPlatform.Flexberry.ORM.ODataService/Model/DataObjectEdmModel.cs
@@ -20,14 +20,22 @@
     using Microsoft.Practices.Unity;
     using Microsoft.Practices.Unity.Configuration;
     using System.Web.OData;
+    using ICSSoft.Services;
 
     /// <summary>
     /// EDM-модель, которая строится на основе сборок с объектами данных (унаследованными от <see cref="DataObject"/>).
     /// </summary>
     public class DataObjectEdmModel : EdmModel
     {
+        /// <summary>
+        /// Service to export data from ORM.
+        /// </summary>
         public IExportService ExportService { get; set; }
 
+        /// <summary>
+        /// Service to export data from ORM.
+        /// </summary>
+        public IODataExportService ODataExportService { get; set; }
         /// <summary>
         /// Ссылка на IDataObjectEdmModelBuilder.
         /// </summary>
@@ -81,12 +89,28 @@
         {
             Contract.Requires<ArgumentNullException>(metadata != null);
             EdmModelBuilder = edmModelBuilder;
-            var container = new UnityContainer();
+            var container = UnityFactory.GetContainer();
             if (container != null)
             {
-                container.LoadConfiguration();
                 if (container.IsRegistered<IExportService>("Export"))
+                {
                     ExportService = container.Resolve<IExportService>("Export");
+                }
+
+                if (container.IsRegistered<IExportService>())
+                {
+                    ExportService = container.Resolve<IExportService>();
+                }
+
+                if (container.IsRegistered<IODataExportService>("Export"))
+                {
+                    ODataExportService = container.Resolve<IODataExportService>("Export");
+                }
+
+                if (container.IsRegistered<IODataExportService>())
+                {
+                    ODataExportService = container.Resolve<IODataExportService>();
+                }
             }
 
             _metadata = metadata;